Divide individual `@connect` from whole-schema validations
