+++ apollo-federation/src/sources/connect/expand/visitors/mod.rs
@@ -115,7 +115,7 @@ where
     /// Walk through the `Group`, visiting each output key. If at any point, one of the
     /// visitor methods returns an error, then the walk will be stopped and the error will be
     /// returned.
-    fn walk(mut self, entry: Group) -> Result<(), <Self as FieldVisitor<Field>>::Error> {
+    fn walk(mut self, entry: Group) -> Result<Self, <Self as FieldVisitor<Field>>::Error> {
         // Start visiting each of the fields
         let mut to_visit =
             VecDeque::from_iter(self.enter_group(&entry)?.into_iter().map(|n| (0i32, n)));
@@ -148,7 +148,7 @@ where
             self.exit_group()?;
         }
 
-        Ok(())
+        Ok(self)
     }
 }
