+++ apollo-federation/src/sources/connect/validation/graphql.rs
@@ -11,8 +11,11 @@ mod strings;
 
 pub(super) use strings::GraphQLString;
 
+use crate::sources::connect::ConnectSpec;
+
 pub(super) struct SchemaInfo<'schema> {
     pub(crate) schema: &'schema Schema,
+    pub(crate) connect_spec: ConnectSpec,
     len: usize,
     lookup: LineColLookup<'schema>,
     pub(crate) connect_directive_name: &'schema Name,
@@ -24,12 +27,14 @@ pub(super) struct SchemaInfo<'schema> {
 impl<'schema> SchemaInfo<'schema> {
     pub(crate) fn new(
         schema: &'schema Schema,
+        connect_spec: ConnectSpec,
         src: &'schema str,
         connect_directive_name: &'schema Name,
         source_directive_name: &'schema Name,
     ) -> Self {
         Self {
             schema,
+            connect_spec,
             len: src.len(),
             lookup: LineColLookup::new(src),
             connect_directive_name,
@@ -73,7 +78,7 @@ mod tests {
         let schema = Schema::parse(src, "testSchema").unwrap();
 
         let name = "unused".try_into().unwrap();
-        let schema_info = SchemaInfo::new(&schema, src, &name, &name);
+        let schema_info = SchemaInfo::new(&schema, ConnectSpec::V0_1, src, &name, &name);
 
         assert_eq!(schema_info.line_col(0), Some((1, 1)));
         assert_eq!(schema_info.line_col(4), Some((2, 4)));
