+++ apollo-federation/src/sources/connect/validation/mod.rs
@@ -14,13 +14,12 @@
     )
 )]
 
+mod connect;
 mod coordinates;
-mod entity;
 mod expression;
-mod extended_type;
 mod graphql;
 mod http;
-mod selection;
+mod schema;
 mod source;
 mod variable;
 
@@ -30,42 +29,26 @@ use std::ops::Range;
 use apollo_compiler::Name;
 use apollo_compiler::Node;
 use apollo_compiler::Schema;
-use apollo_compiler::ast::OperationType;
 use apollo_compiler::ast::Value;
-use apollo_compiler::collections::IndexSet;
-use apollo_compiler::executable::FieldSet;
 use apollo_compiler::name;
 use apollo_compiler::parser::LineColumn;
-use apollo_compiler::parser::Parser;
-use apollo_compiler::schema::Component;
-use apollo_compiler::schema::Directive;
-use apollo_compiler::schema::ExtendedType;
-use apollo_compiler::schema::ObjectType;
 use apollo_compiler::schema::SchemaBuilder;
-use apollo_compiler::validation::Valid;
-use entity::EntityKeyChecker;
-use entity::field_set_error;
-use extended_type::validate_extended_type;
 use itertools::Itertools;
 use strum::IntoEnumIterator;
 use strum_macros::Display;
 use strum_macros::IntoStaticStr;
 use url::Url;
 
-use super::Connector;
 use crate::link::Import;
 use crate::link::Link;
-use crate::link::federation_spec_definition::FEDERATION_FIELDS_ARGUMENT_NAME;
-use crate::link::federation_spec_definition::FEDERATION_KEY_DIRECTIVE_NAME_IN_SPEC;
-use crate::link::federation_spec_definition::FEDERATION_RESOLVABLE_ARGUMENT_NAME;
 use crate::link::spec::Identity;
 use crate::sources::connect::ConnectSpec;
 use crate::sources::connect::spec::schema::SOURCE_DIRECTIVE_NAME_IN_SPEC;
+use crate::sources::connect::validation::connect::fields_seen_by_all_connects;
 use crate::sources::connect::validation::graphql::GraphQLString;
 use crate::sources::connect::validation::graphql::SchemaInfo;
 use crate::sources::connect::validation::source::SourceDirective;
 use crate::subgraph::spec::CONTEXT_DIRECTIVE_NAME;
-use crate::subgraph::spec::EXTERNAL_DIRECTIVE_NAME;
 use crate::subgraph::spec::FROM_CONTEXT_DIRECTIVE_NAME;
 
 // The result of a validation pass on a subgraph
@@ -103,36 +86,34 @@ pub fn validate(source_text: &str, file_name: &str) -> ValidationResult {
         };
     };
 
-    if let Err(err) = ConnectSpec::try_from(&link.url.version) {
-        let available_versions = ConnectSpec::iter().map(ConnectSpec::as_str).collect_vec();
-        let message = if available_versions.len() == 1 {
-            // TODO: No need to branch here once multiple spec versions are available
-            format!("{err}; should be {version}.", version = ConnectSpec::V0_1)
-        } else {
-            // This won't happen today, but it's prepping for 0.2 so we don't forget
-            format!(
-                "{err}; should be one of {available_versions}.",
-                available_versions = available_versions.join(", "),
-            )
-        };
-        return ValidationResult {
-            errors: vec![Message {
-                code: Code::UnknownConnectorsVersion,
-                message,
-                locations: link_directive
-                    .line_column_range(&schema.sources)
-                    .into_iter()
-                    .collect(),
-            }],
-            has_connectors: true,
-            schema,
-        };
-    }
-
-    let federation = Link::for_identity(&schema, &Identity::federation_identity());
-    let external_directive_name = federation
-        .map(|(link, _)| link.directive_name_in_schema(&EXTERNAL_DIRECTIVE_NAME))
-        .unwrap_or(EXTERNAL_DIRECTIVE_NAME.clone());
+    let spec = match ConnectSpec::try_from(&link.url.version) {
+        Err(err) => {
+            let available_versions = ConnectSpec::iter().map(ConnectSpec::as_str).collect_vec();
+            let message = if available_versions.len() == 1 {
+                // TODO: No need to branch here once multiple spec versions are available
+                format!("{err}; should be {version}.", version = ConnectSpec::V0_1)
+            } else {
+                // This won't happen today, but it's prepping for 0.2 so we don't forget
+                format!(
+                    "{err}; should be one of {available_versions}.",
+                    available_versions = available_versions.join(", "),
+                )
+            };
+            return ValidationResult {
+                errors: vec![Message {
+                    code: Code::UnknownConnectorsVersion,
+                    message,
+                    locations: link_directive
+                        .line_column_range(&schema.sources)
+                        .into_iter()
+                        .collect(),
+                }],
+                has_connectors: true,
+                schema,
+            };
+        }
+        Ok(spec) => spec,
+    };
 
     let mut messages = check_conflicting_directives(&schema);
 
@@ -140,6 +121,7 @@ pub fn validate(source_text: &str, file_name: &str) -> ValidationResult {
     let connect_directive_name = ConnectSpec::connect_directive_name(&link);
     let schema_info = SchemaInfo::new(
         &schema,
+        spec,
         source_text,
         &connect_directive_name,
         &source_directive_name,
@@ -152,17 +134,19 @@ pub fn validate(source_text: &str, file_name: &str) -> ValidationResult {
         .map(|directive| directive.name)
         .collect_vec();
 
-    let mut seen_fields = IndexSet::default();
-
-    let connect_errors = schema.types.values().flat_map(|extended_type| {
-        validate_extended_type(
-            extended_type,
-            &schema_info,
-            &all_source_names,
-            &mut seen_fields,
-        )
-    });
-    messages.extend(connect_errors);
+    match fields_seen_by_all_connects(&schema_info, &all_source_names) {
+        Ok(fields_seen_by_connectors) => {
+            // Don't run schema-wide checks if any connectors failed to validate
+            messages.extend(schema::validate(
+                &schema_info,
+                file_name,
+                fields_seen_by_connectors,
+            ))
+        }
+        Err(errs) => {
+            messages.extend(errs);
+        }
+    }
 
     if source_directive_name == DEFAULT_SOURCE_DIRECTIVE_NAME
         && messages
@@ -178,136 +162,11 @@ pub fn validate(source_text: &str, file_name: &str) -> ValidationResult {
         });
     }
 
-    if should_do_advanced_validations(&messages) {
-        messages.extend(check_seen_fields(
-            &schema_info,
-            &seen_fields,
-            &external_directive_name,
-        ));
-
-        match advanced_validations(&schema, file_name, ConnectSpec::V0_1) {
-            Ok(multiple) => messages.extend(multiple),
-            Err(Some(single)) => messages.push(single),
-            _ => {} // let the rest of composition handle this
-        };
-    }
-
     ValidationResult {
         errors: messages,
         has_connectors: true,
         schema,
     }
-}
-
-fn advanced_validations(
-    schema: &Schema,
-    subgraph_name: &str,
-    spec: ConnectSpec,
-) -> Result<Vec<Message>, Option<Message>> {
-    let mut messages = vec![];
-
-    let connectors = Connector::from_schema(schema, subgraph_name, spec).map_err(|_| None)?;
-
-    let mut entity_checker = EntityKeyChecker::default();
-
-    for (field_set, directive) in find_all_resolvable_keys(schema) {
-        entity_checker.add_key(&field_set, directive);
-    }
-
-    for (_, connector) in connectors {
-        if let Some(field_set) = connector.resolvable_key(schema).map_err(|_| {
-            let variables = connector.variable_references().collect_vec();
-            field_set_error(&variables, connector.id.directive.field.type_name())
-        })? {
-            entity_checker.add_connector(field_set);
-        }
-    }
-
-    messages.extend(entity_checker.check_for_missing_entity_connectors(schema));
-
-    Ok(messages)
-}
-
-/// We'll avoid doing this work if there are bigger issues with the schema.
-/// Otherwise we might emit a large number of diagnostics that will
-/// distract from the main problems.
-fn should_do_advanced_validations(messages: &[Message]) -> bool {
-    messages.is_empty()
-}
-
-/// Check that all fields defined in the schema are resolved by a connector.
-fn check_seen_fields(
-    schema: &SchemaInfo,
-    seen_fields: &IndexSet<(Name, Name)>,
-    external_directive_name: &Name,
-) -> Vec<Message> {
-    let all_fields: IndexSet<_> = schema
-        .types
-        .values()
-        .filter_map(|extended_type| {
-            if extended_type.is_built_in() {
-                return None;
-            }
-            // ignore root fields, we have different validations for them
-            if schema.root_operation(OperationType::Query) == Some(extended_type.name())
-                || schema.root_operation(OperationType::Mutation) == Some(extended_type.name())
-                || schema.root_operation(OperationType::Subscription) == Some(extended_type.name())
-            {
-                return None;
-            }
-            let coord = |(name, _): (&Name, _)| (extended_type.name().clone(), name.clone());
-
-            // ignore all fields on objects marked @external
-            if extended_type
-                .directives()
-                .iter()
-                .any(|dir| &dir.name == external_directive_name)
-            {
-                return None;
-            }
-
-            match extended_type {
-                ExtendedType::Object(object) => {
-                    // ignore fields marked @external
-                    Some(
-                        object
-                            .fields
-                            .iter()
-                            .filter(|(_, def)| {
-                                !def.directives
-                                    .iter()
-                                    .any(|dir| &dir.name == external_directive_name)
-                            })
-                            .map(coord),
-                    )
-                }
-                ExtendedType::Interface(_) => None, // TODO: when interfaces are supported (probably should include fields from implementing/member types as well)
-                _ => None,
-            }
-        })
-        .flatten()
-        .collect();
-
-    (&all_fields - seen_fields).iter().map(move |(parent_type, field_name)| {
-        let Ok(field_def) = schema.type_field(parent_type, field_name) else {
-            // This should never happen, but if it does, we don't want to panic
-            return Message {
-                code: Code::GraphQLError,
-                message: format!(
-                    "Field `{parent_type}.{field_name}` is missing from the schema.",
-                ),
-                locations: Vec::new(),
-            };
-        };
-        Message {
-            code: Code::ConnectorsUnresolvedField,
-            message: format!(
-                "No connector resolves field `{parent_type}.{field_name}`. It must have a `@{connect_directive_name}` directive or appear in `@{connect_directive_name}(selection:)`.",
-                connect_directive_name = schema.connect_directive_name
-            ),
-            locations: field_def.line_column_range(&schema.sources).into_iter().collect(),
-        }
-    }).collect()
 }
 
 fn check_conflicting_directives(schema: &Schema) -> Vec<Message> {
@@ -376,60 +235,6 @@ fn parse_url<Coordinate: Display + Copy>(
             .collect(),
     })?;
     http::url::validate_base_url(&url, coordinate, value, str_value, schema)
-}
-
-/// For an object type, get all the keys (and directive nodes) that are resolvable.
-///
-/// The FieldSet returned here is what goes in the `fields` argument, so `id` in `@key(fields: "id")`
-fn resolvable_key_fields<'a>(
-    object: &'a Node<ObjectType>,
-    schema: &'a Schema,
-) -> impl Iterator<Item = (FieldSet, &'a Component<Directive>)> {
-    object
-        .directives
-        .iter()
-        .filter(|directive| directive.name == FEDERATION_KEY_DIRECTIVE_NAME_IN_SPEC)
-        .filter(|directive| {
-            directive
-                .arguments
-                .iter()
-                .find(|arg| arg.name == FEDERATION_RESOLVABLE_ARGUMENT_NAME)
-                .and_then(|arg| arg.value.to_bool())
-                .unwrap_or(true)
-        })
-        .filter_map(|directive| {
-            if let Some(fields_str) = directive
-                .arguments
-                .iter()
-                .find(|arg| arg.name == FEDERATION_FIELDS_ARGUMENT_NAME)
-                .map(|arg| &arg.value)
-                .and_then(|value| value.as_str())
-            {
-                Parser::new()
-                    .parse_field_set(
-                        Valid::assume_valid_ref(schema),
-                        object.name.clone(),
-                        fields_str.to_string(),
-                        "",
-                    )
-                    .ok()
-                    .map(|field_set| (field_set, directive))
-            } else {
-                None
-            }
-        })
-}
-
-fn find_all_resolvable_keys(schema: &Schema) -> Vec<(FieldSet, &Component<Directive>)> {
-    schema
-        .types
-        .values()
-        .flat_map(|extended_type| match extended_type {
-            ExtendedType::Object(object) => Some(resolvable_key_fields(object, schema)),
-            _ => None,
-        })
-        .flatten()
-        .collect()
 }
 
 type DirectiveName = Name;
